Flytter vedtaksfarger inn i utils
Fikser også oppsummeringstabell i testen
