+++ src/frontend/pages/fagsak/vedtak/utils.ts
@@ -1,3 +1,4 @@
+import { TagProps } from '@navikt/ds-react';
 import type { VedtaksbrevFormData } from './schema';
 import type {
     PakrevdBegrunnelseUpdateItem,
@@ -6,9 +7,16 @@ import type {
     VedtaksbrevData,
     VedtaksbrevDataWritable,
     VedtaksbrevRedigerbareDataUpdate,
+    Vedtaksresultat,
 } from '~/generated-new';
 
 type RentekstElement = { type: 'rentekst'; tekst: string };
+
+export const vedtaksresultatFarger: Record<Vedtaksresultat, TagProps['data-color']> = {
+    DelvisTilbakebetaling: 'meta-purple',
+    IngenTilbakebetaling: 'brand-beige',
+    FullTilbakebetaling: 'meta-lime',
+};
 
 export const elementArrayTilTekst = (
     elementer: readonly { type: string; tekst?: string }[]
